Created a CreateJobPost component which will be displayed when company user wants to add a new job post. Modified CompanyJobs to show the jobs the company user has posted.
